requirements.txt
Django==5.0
djangorestframework==3.15
psycopg2-binary==2.9.10
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
gunicorn==22.0.0